refactor(lix): give shared row contracts engine ownership
