+++ packages/lix/src/row_columnar.rs
@@ -1,9 +1,12 @@
 //! Row-specific contract layered over generic immutable columnar row groups.
 
 use std::collections::BTreeMap;
+use std::ops::Deref;
 
 use crate::changelog::CommitId;
-use crate::columnar_row_group::{RowGroupRowLocation, RowGroupSetId};
+use crate::columnar_row_group::{
+    EncodedRowGroupSet, ROW_GROUP_MAX_ROWS, RowGroupRowLocation, RowGroupSetId,
+};
 
 pub(crate) const ROW_COLUMNAR_LOSSLESS_SNAPSHOT_METADATA_KEY: &str =
     "lix.row_columnar.lossless_snapshot.v1";
@@ -26,7 +29,7 @@ pub(crate) fn row_identity_column_index(
 }
 
 pub(crate) struct RowColumnarWriteSets {
-    sets: BTreeMap<(CommitId, String), crate::columnar_row_group::EncodedRowGroupSet>,
+    sets: BTreeMap<(CommitId, String), EncodedRowGroupSet>,
     state_row_locations: StateRowLocations,
 }
 
@@ -65,25 +68,15 @@ impl RowColumnarWriteSets {
         }
     }
 
-    pub(crate) fn get(
-        &self,
-        key: &(CommitId, String),
-    ) -> Option<&crate::columnar_row_group::EncodedRowGroupSet> {
+    pub(crate) fn get(&self, key: &(CommitId, String)) -> Option<&EncodedRowGroupSet> {
         self.sets.get(key)
     }
 
-    pub(crate) fn take(
-        &mut self,
-        key: &(CommitId, String),
-    ) -> Option<crate::columnar_row_group::EncodedRowGroupSet> {
+    pub(crate) fn take(&mut self, key: &(CommitId, String)) -> Option<EncodedRowGroupSet> {
         self.sets.remove(key)
     }
 
-    pub(crate) fn insert(
-        &mut self,
-        key: (CommitId, String),
-        value: crate::columnar_row_group::EncodedRowGroupSet,
-    ) {
+    pub(crate) fn insert(&mut self, key: (CommitId, String), value: EncodedRowGroupSet) {
         self.sets.insert(key, value);
     }
 
@@ -107,14 +100,8 @@ impl RowColumnarWriteSets {
             StateRowLocations::None => None,
             StateRowLocations::Dense { row_count } if state_row_index < *row_count => {
                 Some(RowGroupRowLocation {
-                    group_index: u32::try_from(
-                        state_row_index / crate::columnar_row_group::ROW_GROUP_MAX_ROWS,
-                    )
-                    .ok()?,
-                    row_index: u32::try_from(
-                        state_row_index % crate::columnar_row_group::ROW_GROUP_MAX_ROWS,
-                    )
-                    .ok()?,
+                    group_index: u32::try_from(state_row_index / ROW_GROUP_MAX_ROWS).ok()?,
+                    row_index: u32::try_from(state_row_index % ROW_GROUP_MAX_ROWS).ok()?,
                 })
             }
             StateRowLocations::Dense { .. } => None,
@@ -134,4 +121,70 @@ pub(crate) fn row_group_set_id(commit_id: CommitId, schema_key: &str) -> RowGrou
     let mut id = [0_u8; 16];
     id.copy_from_slice(&digest.finalize().as_bytes()[..16]);
     RowGroupSetId::new(id)
+}
+
+#[derive(Clone, Debug)]
+pub(crate) struct EncodedRowGroups {
+    pub(crate) encoded: EncodedRowGroupSet,
+    pub(crate) input_locations: RowGroupLocations,
+}
+
+/// Input-row to physical-row mapping for one sealed row generation.
+///
+/// Identity-preserving batches use arithmetic coordinates and retain no
+/// row-cardinal location column. Clustered layouts keep the explicit
+/// permutation required to map their reordered rows back to statement order.
+#[derive(Clone, Debug)]
+pub(crate) enum RowGroupLocations {
+    Dense { row_count: usize },
+    Explicit(Vec<RowGroupRowLocation>),
+}
+
+impl RowGroupLocations {
+    pub(crate) fn len(&self) -> usize {
+        match self {
+            Self::Dense { row_count } => *row_count,
+            Self::Explicit(locations) => locations.len(),
+        }
+    }
+
+    pub(crate) fn location(&self, input_index: usize) -> Option<RowGroupRowLocation> {
+        match self {
+            Self::Dense { row_count } if input_index < *row_count => Some(RowGroupRowLocation {
+                group_index: u32::try_from(input_index / ROW_GROUP_MAX_ROWS).ok()?,
+                row_index: u32::try_from(input_index % ROW_GROUP_MAX_ROWS).ok()?,
+            }),
+            Self::Dense { .. } => None,
+            Self::Explicit(locations) => locations.get(input_index).copied(),
+        }
+    }
+
+    pub(crate) fn iter(&self) -> impl ExactSizeIterator<Item = RowGroupRowLocation> + '_ {
+        (0..self.len()).map(|input_index| {
+            self.location(input_index)
+                .expect("row-group location covers every input row")
+        })
+    }
+}
+
+impl PartialEq for RowGroupLocations {
+    fn eq(&self, other: &Self) -> bool {
+        self.len() == other.len() && self.iter().eq(other.iter())
+    }
+}
+
+impl Eq for RowGroupLocations {}
+
+impl Deref for EncodedRowGroups {
+    type Target = EncodedRowGroupSet;
+
+    fn deref(&self) -> &Self::Target {
+        &self.encoded
+    }
+}
+
+impl EncodedRowGroups {
+    pub(crate) fn into_parts(self) -> (EncodedRowGroupSet, RowGroupLocations) {
+        (self.encoded, self.input_locations)
+    }
 }
