+++ packages/lix/src/transaction_types.rs
@@ -14,10 +14,11 @@ use crate::catalog::SchemaPlanId;
 use crate::changelog::{ChangeId, CommitId};
 use crate::common::{LixTimestamp, MutationIdentity, RequestBlobSpliceProvenance, SharedStr};
 use crate::functions::FunctionProviderHandle;
-use crate::hot_state::{CertifiedCurrentStatePredecessor, MaterializedHotStateRow};
 use crate::json_store::JsonRef;
-use crate::plugin::runtime::{WasmCertifiedRowBatch, WasmTypedRow};
+use crate::row_payload::CertifiedRowBatch as WasmCertifiedRowBatch;
+use crate::row_payload::TypedRow as WasmTypedRow;
 use crate::row_pk::RowPk;
+use crate::row_state::CertifiedCurrentStatePredecessor;
 use crate::tracked_state::OrderedAddressableCommitDeltaStage;
 use bytes::Bytes;
 use lix_schema::Jsonb;
@@ -549,7 +550,7 @@ pub(crate) struct CertifiedParameterBatch {
     branch_id: SharedStr,
     untracked: bool,
     certificate: CertifiedRawWriteBatchPreparation,
-    row_columnar: Option<crate::sql2::EncodedRowGroups>,
+    row_columnar: Option<crate::row_columnar::EncodedRowGroups>,
 }
 
 struct CertifiedSnapshots {
@@ -681,7 +682,10 @@ impl CertifiedParameterBatch {
         })
     }
 
-    pub(crate) fn with_row_columnar(mut self, row_columnar: crate::sql2::EncodedRowGroups) -> Self {
+    pub(crate) fn with_row_columnar(
+        mut self,
+        row_columnar: crate::row_columnar::EncodedRowGroups,
+    ) -> Self {
         self.row_columnar = Some(row_columnar);
         self
     }
@@ -3155,7 +3159,7 @@ pub(crate) struct StagedIndexRow {
     /// not extraction found a value, so dropping the `None` entries would
     /// silently narrow witness coverage and leave a column permanently
     /// unwitnessed — a slow read, not a wrong one, but an invisible one.
-    pub(crate) columns: Vec<(u16, Option<crate::hot_state::HotIndexValue>)>,
+    pub(crate) columns: Vec<(u16, Option<crate::row_state::HotIndexValue>)>,
 }
 
 /// Everything the commit-time hot index hook needs, produced by validation.
@@ -3269,7 +3273,7 @@ struct DenseCertifiedParameterSlots {
     direct_change_ids: Option<OrderedAddressableCommitDeltaStage>,
     /// Frontend-built row groups over the same certified typed columns.
     /// Topology-changing operations drop this derived accelerator.
-    row_columnar: Option<crate::sql2::EncodedRowGroups>,
+    row_columnar: Option<crate::row_columnar::EncodedRowGroups>,
     /// Exact authenticated predecessor evidence is row-aligned with the
     /// certified identity column. Ordinary replacement batches therefore
     /// retain their compact representation through publication.
@@ -4231,7 +4235,7 @@ impl PreparedStateBatch {
 
     pub(crate) fn take_dense_row_columnar(
         &mut self,
-    ) -> Option<(CommitId, String, crate::sql2::EncodedRowGroups)> {
+    ) -> Option<(CommitId, String, crate::row_columnar::EncodedRowGroups)> {
         let dense = self.dense_certified_parameter.as_mut()?;
         let commit_id = dense.commit_id?;
         let encoded = dense.row_columnar.take()?;
@@ -4530,108 +4534,55 @@ pub(crate) fn materialize_jsonb_shared(value: &Jsonb) -> SharedStr {
     )
 }
 
-impl From<PreparedStateRowRef<'_>> for MaterializedHotStateRow {
-    fn from(row: PreparedStateRowRef<'_>) -> Self {
-        Self {
-            row_pk: row.row_pk.clone(),
-            schema_key: row.schema_key.to_string(),
-            file_id: row.file_id.map(ToString::to_string),
-            snapshot_content: None,
-            metadata: row.metadata.map(materialize_jsonb_shared),
-            deleted: row.snapshot.is_none(),
-            created_at: row.created_at,
-            updated_at: row.updated_at,
-            global: row.global,
-            change_id: row.change_id,
-            commit_id: row.commit_id,
-            untracked: row.untracked,
-            branch_id: Arc::from(row.branch_id.as_str()),
-        }
+pub(crate) fn duplicate_insert_identity_message(
+    schema_key: &str,
+    _row_pk: &RowPk,
+    branch_id: Option<&str>,
+    origin: Option<&TransactionWriteOrigin>,
+) -> String {
+    if let Some(message) = logical_primary_key_violation_message(origin) {
+        return message;
+    }
+    match branch_id {
+        Some(branch_id) => format!(
+            "primary-key constraint violation on schema '{schema_key}': INSERT would duplicate a primary key in branch '{branch_id}'"
+        ),
+        None => format!(
+            "primary-key constraint violation on schema '{schema_key}': INSERT would duplicate a primary key"
+        ),
     }
 }
 
-/// Builds the legacy owned DTO at an explicit JSON projection boundary.
-/// Durable and batch-native paths keep the typed sidecar instead; persistent
-/// filesystem indexes currently own this DTO and therefore request the
-/// transient projection deliberately.
-pub(crate) fn materialized_hot_state_row_with_snapshot_projection(
-    row: PreparedStateRowRef<'_>,
-) -> Result<MaterializedHotStateRow, LixError> {
-    let mut materialized = MaterializedHotStateRow::from(row);
-    materialized.snapshot_content = row
-        .snapshot
-        .map(|payload| {
-            WasmTypedRow::decode_durable_payload(
-                Arc::from(payload),
-                row.schema_key.as_str(),
-                row.row_pk,
-            )?
-            .to_json_shared()
+pub(crate) fn logical_primary_key_violation_message(
+    origin: Option<&TransactionWriteOrigin>,
+) -> Option<String> {
+    let origin = origin?;
+    if origin.operation != TransactionWriteOperation::Insert {
+        return None;
+    }
+    let primary_key = origin.primary_key.as_ref()?;
+    Some(format!(
+        "primary-key constraint violation on table '{}': INSERT would duplicate {}",
+        origin.surface,
+        format_logical_primary_key(primary_key)
+    ))
+}
+
+fn format_logical_primary_key(primary_key: &LogicalPrimaryKey) -> String {
+    primary_key
+        .columns
+        .iter()
+        .enumerate()
+        .map(|(index, column)| {
+            let value = primary_key
+                .values
+                .get(index)
+                .map(String::as_str)
+                .unwrap_or("<missing>");
+            format!("{column} '{value}'")
         })
-        .transpose()?;
-    Ok(materialized)
-}
-
-#[cfg(test)]
-impl From<TestPreparedStateRow> for MaterializedHotStateRow {
-    fn from(row: TestPreparedStateRow) -> Self {
-        let deleted = row.snapshot.is_none();
-        let snapshot_content = row.snapshot.as_deref().map(|snapshot| {
-            WasmTypedRow::decode_durable_payload(
-                Arc::from(snapshot),
-                row.schema_key.as_str(),
-                &row.row_pk,
-            )
-            .and_then(|typed| typed.to_json_shared())
-            .expect("test prepared snapshot should decode")
-        });
-        Self {
-            row_pk: row.row_pk,
-            schema_key: row.schema_key.into(),
-            file_id: row.file_id.map(Into::into),
-            snapshot_content,
-            metadata: row
-                .metadata
-                .map(|metadata| materialize_jsonb_shared(&metadata)),
-            deleted,
-            created_at: row.created_at,
-            updated_at: row.updated_at,
-            global: row.global,
-            change_id: row.change_id,
-            commit_id: row.commit_id,
-            untracked: row.untracked,
-            branch_id: Arc::from(row.branch_id.as_str()),
-        }
-    }
-}
-
-#[cfg(test)]
-impl From<&TestPreparedStateRow> for MaterializedHotStateRow {
-    fn from(row: &TestPreparedStateRow) -> Self {
-        Self {
-            row_pk: row.row_pk.clone(),
-            schema_key: row.schema_key.to_string(),
-            file_id: row.file_id.as_ref().map(ToString::to_string),
-            snapshot_content: row.snapshot.as_deref().map(|snapshot| {
-                WasmTypedRow::decode_durable_payload(
-                    Arc::from(snapshot),
-                    row.schema_key.as_str(),
-                    &row.row_pk,
-                )
-                .and_then(|typed| typed.to_json_shared())
-                .expect("test prepared snapshot should decode")
-            }),
-            metadata: row.metadata.as_ref().map(materialize_jsonb_shared),
-            deleted: row.snapshot.is_none(),
-            created_at: row.created_at,
-            updated_at: row.updated_at,
-            global: row.global,
-            change_id: row.change_id,
-            commit_id: row.commit_id,
-            untracked: row.untracked,
-            branch_id: Arc::from(row.branch_id.as_str()),
-        }
-    }
+        .collect::<Vec<_>>()
+        .join(", ")
 }
 
 #[cfg(test)]
@@ -5181,7 +5132,6 @@ mod tests {
         );
     }
 
-
     #[test]
     fn prepared_batch_compacts_superseded_owner_columns() {
         let origin_key: SharedStr = "one-execution".into();
@@ -5276,7 +5226,6 @@ mod tests {
         assert_eq!(second_key.values, ["file-b"]);
     }
 
-
     #[test]
     fn ten_thousand_write_row_clones_retain_identifier_buffers() {
         let schema_key = SharedStr::from("bulk_schema");
@@ -5368,10 +5317,6 @@ mod tests {
         assert!(!staged.is_inline());
         assert_eq!(staged.json_ref, expected);
     }
-
-
-
-
 
     #[test]
     fn certified_transaction_rows_release_without_moving_native_columns() {
@@ -5470,8 +5415,6 @@ mod tests {
         assert!(!staged.retains_decoded_value_for_tests());
         assert_eq!(staged.normalized(), r#"{"id":"row-1"}"#);
     }
-
-
 
     #[test]
     fn decoded_sql_rows_canonicalize_into_one_exact_batch_arena() {
